+++ pyrad/lbl/hitran/isotopologues.py
@@ -1,13 +1,14 @@
 from collections import namedtuple
 from html.parser import HTMLParser
+from urllib.request import urlopen
 
 from numpy import float32
 
 
-Isotopologue = namedtuple("Isotopologue", ["id", "abundance", "mass"])
+Isotopologue = namedtuple("Isotopologue", ["abundance", "id", "mass"])
 
 
-class HitranIsotopologueHTMLParser(HTMLParser):
+class _Parser(HTMLParser):
     """Parses the isotopologue properties for a molecule from the HITRAN website.
 
     Attributes:
@@ -61,7 +62,7 @@ class HitranIsotopologueHTMLParser(HTMLParser):
         elif tag == "tr" and self.inside_row and self.column > 0:
             abundance = self.abundance.replace("\xa0{}\xa010".format(chr(215)), "e")
             self.abundance = ""
-            self.iso_list.append(Isotopologue(int(self.id), abundance, float32(self.mass)))
+            self.iso_list.append(Isotopologue(id=int(self.id), abundance=abundance, mass=float32(self.mass)))
             self.id = ""
             self.mass = ""
             self.inside_row = False
@@ -86,61 +87,7 @@ class HitranIsotopologueHTMLParser(HTMLParser):
             self.column += 1
 
 
-class HitranMoleculeIdHTMLParser(HTMLParser):
-    """Parses HITRAN molecule ids from the HITRAN website.
-
-    Attributes:
-        formula: Chemical formula of the current molecule.
-        id: HITRAN integer identifier of the current molecule.
-        inside_row: Flag telling if the parser is inside a row of a HTML table.
-        inside_table: Flag telling if the parser is inside the correct HTML table.
-        molecule_ids: Dictionary mapping chemical formulae to their HITRAN ids.
-    """
-    def __init__(self):
-        HTMLParser.__init__(self)
-        self.formula = ""
-        self.id = ""
-        self.inside_row = False
-        self.inside_table = False
-        self.molecule_ids = {}
-
-    def handle_data(self, data):
-        """Stores the current HITRAN id or chemical symbol, based on the parser's current state.
-
-        Args:
-            data: The data contained in the current HTML tag.
-        """
-        if self.inside_row:
-            if self.column == 1:
-                self.id += data.strip()
-            elif self.column == 2:
-                self.formula += data.strip()
-
-    def handle_endtag(self, tag):
-        """Resets flags and stores data based on the parser's current state.
-
-        Args:
-           tag: Current HTML tag.
-        """
-        if tag == "table" and self.inside_table:
-            self.inside_table = False
-        if tag == "tr" and self.inside_row and self.column > 0:
-            self.molecule_ids[self.formula] = int(self.id)
-            self.formula = ""
-            self.id = ""
-            self.inside_row = False
-
-    def handle_starttag(self, tag, attrs):
-        """Sets flags based on the parser's current state.
-
-        Args:
-            tag: Current HTML tag.
-            attrs: List of tuples containing HTML tag attributes.
-        """
-        if tag == "table" and ("class", "list-table") in attrs:
-            self.inside_table = True
-        elif tag == "tr" and self.inside_table:
-            self.column = 0
-            self.inside_row = True
-        elif tag == "td" and self.inside_row:
-            self.column += 1
+def isotopologues(url):
+    parser = _Parser()
+    parser.feed(urlopen(url).read().decode("utf-8"))
+    return parser.isotopologues
